+++ packages/ente-ui/src/routes/Slots.tsx
@@ -35,6 +35,7 @@ const lang = createTranslation({
       date: "Date",
       from: "From",
       to: "To",
+      forSchool: "For School",
       signed: "Signed",
       teacher: "Teacher"
     },
@@ -48,6 +49,7 @@ const lang = createTranslation({
       date: "Datum",
       from: "Von",
       to: "Bis",
+      forSchool: "Schulisch",
       signed: "Unterschrieben",
       teacher: "Lehrer"
     },
@@ -101,6 +103,7 @@ export class Slots extends React.Component<SlotsProps> {
           lang.headers.date,
           lang.headers.from,
           lang.headers.to,
+          lang.headers.forSchool,
           {
             name: lang.headers.signed,
             options: {
@@ -118,6 +121,7 @@ export class Slots extends React.Component<SlotsProps> {
           slot.get("date").toLocaleDateString(),
           "" + slot.get("from"),
           "" + slot.get("to"),
+          slot.get("forSchool") ? lang.yes : lang.no,
           slot.get("signed") ? lang.yes : lang.no,
           Maybe.fromNull(slot.get("teacherId")).cata(
             () => lang.deleted,